requirements.txt
streamlit
python-dotenv
openai
google-generativeai
langchain
langchain-community
langchain-google-genai
PyPDF2
pytesseract
Pillow
faiss-cpu
